requirements.txt
# --- Core dependencies ---
typer[all]
llama-cpp-python==0.3.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
pypdf==4.3.1
python-frontmatter==1.1.0
rich==13.7.1
pyyaml==6.0.2
PyPDF2

# --- Optional utilities ---
wheel
setuptools
